requirements.txt
numpy==1.26.4
pygame>=2.5.0
PyOpenGL>=3.1.6
PyOpenGL_accelerate>=3.1.6
opencv-python>=4.8.0
mediapipe>=0.10.0
scipy>=1.11.0
psutil>=5.9.0
picamera2>=0.3.0
